hotel/requirement.txt
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
pytz==2018.9
